split notification logic into a HOC and a UI component
